src/requirements.txt
contractions==0.0.58
emoji==1.4.1
matplotlib==3.3.3
NRCLex==2.0.2
numpy==1.19.2
pandas==1.1.4
pymongo==3.12.0
python_dateutil==2.8.2
requests==2.25.1
scikit_learn==1.1.1
six==1.15.0
transformers==4.9.1
schedule==1.1.0
